MDEV-32188 make TIMESTAMP use whole 32-bit unsigned range
This patch extends the timestamp from
2038-01-19 03:14:07.999999 to 2106-02-07 06:28:15.999999
for 64 bit hardware and OS where 'long' is 64 bits.
This is true for 64 bit Linux but not for Windows.

This is done by treating the 32 bit stored int as unsigned instead of
signed.  This is safe as MariaDB has never accepted dates before the epoch
(1970).
The benefit of this approach that for normal timestamp the storage is
compatible with earlier version.

However for system versioned tables we currently store the 2038 year max
timestamp as the 'max timestamp' which is used to detect current values.
This patch stores the new 2106 max value as the max timestamp.
For now, this handled by treating old rows with the old 2038 timestamp
as it would be the new max 2106 timestamp.
At some point, before 2016, we have to fix mysql_upgrade to update all
system versioning timestamps to use the new 2106 max timestamp.