MDEV-29445 preparation (broken): Add buf_pool_t::is_uncompressed()
This will cause occasional crashes in the InnoDB buffer pool resizing tests
until all related data structures have been refactored and is_uncompressed()
reimplemented.

The idea is that all tests on whether a block descriptor is for a
compressed-only block descriptor of a ROW_FORMAT=COMPRESSED page
will be replaced by buf_pool_t::is_uncompressed(), so that we will
avoid doing busy work in buf_page_t::frame() when the caller is not
interested in the exact value of the pointer.