MDEV-32212  DELETE with ORDER BY and semijoin optimization causing crash
Classes for multitable update aren't currently able to take into
account any filesort structure in read_record->sort_info and assume
that one hasn't been built.

Semijoin optimization on an innocent looking query, such as

DELETE FROM t1 WHERE c1 IN (select c2 from t2) ORDER BY c1;

can cause a filesort structure to be built while forcing use of the
multitable delete classes, previously used only in the context of
multiple-table delete syntax which does not allow an order by clause.

This filesort structure causes the read_record classes in
init_read_record to be one of the rr_unpack_from_buffer templated
routines, none of which track the table->file cursor assumed to be
positioned correctly in multi_delete::send_data.
This triggers an ASSERT in InnoDB and may cause problems with other
storage engines.

We can disable this semijoin optimization with an optimizer switch

set optimizer_switch='semijoin=off';

but this needs to be selective, so in JOIN::prepare, just before we
check for subquery rewrites we check for the condition above

1) We are executing a delete command
2) we have an outer select (i.e there is more than one)
3) the effective outer select_lex has order by list elements

and temporarily disable OPTIMIZER_SWITCH_SEMIJOIN during this
part of the optimization.