MDEV-28509 Dereferenced null pointer of type 'struct JOIN_TAB' in add_key_field
Pushdown conditions created to optimize GROUP BY should correctly determine
if a field relies on an outer table or not.  In the case of this bug, we
incorrectly match a field from an outer table because, while the fields themselves
matched, their corresponding used_table bitmaps did not.  This led to the optimizer
extracting a pushdown condition that referred to an outer table alias that
didn't exist.